Pogo.com/sign in problems
Check for correct User Name or Screen Name. Most of the time we rely on the browser’s auto-complete feature and select wrongly entered screen name in the past.
Passwords are case sensitive. Therefore, always check the letter’s cases of the entered password. If possible, type the password as you use in a notepad and then copy and paste the letters in the password field. Below is a list of things you should always check before entering into troubleshooting.
1. If any of your family member shares the same account, ask them if they might have changed the screen name or password.
2. Look into your email inbox to get the actual screen name when you had registered with Pogo
3. If you find any point relevant to your case, you can go to the article explaining how to reset password of your Pogo account and follow all the steps carefully to reset your pogo account password.

For more detailed information visit: https://pogo-supportcenter.com/2020/04/01/fix-pogo-sign-in-issue/